@@ -3,7 +3,7 @@ name = "udata-hydra"
 version = "2.0.0.dev"
 description = "Async crawler and parsing service for data.gouv.fr"
 authors = [{ name = "Opendata Team", email = "opendatateam@data.gouv.fr" }]
-requires-python = ">=3.9,<3.13"
+requires-python = ">=3.11,<3.13"
 license = { text = "MIT" }
 readme = "README.md"
 
@@ -16,28 +16,27 @@ license = "MIT"
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.11"
+aiocontextvars = "^0.2.2"
 aiohttp = "^3.10.3"
 asyncpg = "^0.29.0"
+coloredlogs = "^15.0.1"
+csv-detective = "0.7.2"
+dateparser = "^1.1.7"
 humanfriendly = "^10.0"
 marshmallow = "^3.14.1"
 minicli = "^0.5.3"
-progressist = "^0.1.0"
+minio = "7.2.7"
+pyarrow = "16.1.0"
+python-dateutil = "^2.8.2"
 python-magic = "^0.4.25"
+progressist = "^0.1.0"
 redis = "^4.1.4"
-sentry-sdk = "^2.10.0"
-aiocontextvars = "^0.2.2"
-coloredlogs = "^15.0.1"
 rq = "^1.11.1"
-toml = "^0.10.2"
+sentry-sdk = "^2.10.0"
+sqlalchemy = "^1.4.46"
 str2float = "^0.0.9"
 str2bool = "^1.1"
-sqlalchemy = "^1.4.46"
-dateparser = "^1.1.7"
-python-dateutil = "^2.8.2"
-csv-detective = "0.7.2.dev800"
-pyarrow = "16.1.0"
-minio = "7.2.7"
 
 [tool.poetry.group.dev.dependencies]
 aiohttp-devtools = "^1.0.post0"
