@@ -1,4 +1,5 @@
 from io import BytesIO
+from typing import Optional
 
 import pyarrow.parquet as pq
 import pytest
@@ -26,7 +27,8 @@ async def test_parquet_conversion(
 ):
     filename, expected_count = file_and_count
     file_path = f"tests/data/{filename}"
-    inspection = await perform_csv_inspection(file_path)
+    inspection: Optional[dict] = await perform_csv_inspection(file_path)
+    assert inspection
     columns = inspection["columns"]
     columns = {
         f"{c}__hydra_renamed" if c.lower() in RESERVED_COLS else c: v["python_type"]
@@ -35,8 +37,7 @@ async def test_parquet_conversion(
     _, table = save_as_parquet(
         records=generate_records(file_path, inspection, columns),
         columns=columns,
-        output_name=None,
-        save_output=False,
+        output_filename=None,
     )
     assert len(table) == expected_count
     fake_file = BytesIO()
