@@ -1,13 +1,16 @@
+import importlib.metadata
+import logging
 import os
+import tomllib
 from pathlib import Path
 
-import toml
+log = logging.getLogger("udata-hydra")
 
 
 class Configurator:
     """Loads a dict of config from TOML file(s) and behaves like an object, ie config.VALUE"""
 
-    configuration = None
+    configuration: dict = {}
 
     def __init__(self):
         if not self.configuration:
@@ -15,15 +18,21 @@ class Configurator:
 
     def configure(self) -> None:
         # load default settings
-        configuration = toml.load(Path(__file__).parent / "config_default.toml")
+        with open(Path(__file__).parent / "config_default.toml", "rb") as f:
+            configuration: dict = tomllib.load(f)
 
         # override with local settings
         local_settings = os.environ.get("HYDRA_SETTINGS", Path.cwd() / "config.toml")
         if Path(local_settings).exists():
-            configuration.update(toml.load(local_settings))
+            with open(Path(local_settings), "rb") as f:
+                configuration.update(tomllib.load(f))
 
         self.configuration = configuration
         self.check()
+
+        # add project metadata to config
+        self.configuration["APP_NAME"] = "udata-hydra"
+        self.configuration["APP_VERSION"] = importlib.metadata.version("udata-hydra")
 
     def override(self, **kwargs) -> None:
         self.configuration.update(kwargs)
