@@ -17,14 +17,13 @@ PYTHON_TYPE_TO_PA = {
 def save_as_parquet(
     records: Generator,
     columns: dict,
-    output_name: str,
-    save_output: bool = True,
+    output_filename: str | None = None,
 ) -> tuple[str, pa.Table]:
-    # the "save_output" argument is only used in tests
+    # the "output_name = None" case is only used in tests
     table = pa.Table.from_pylist(
         [{c: v for c, v in zip(columns, values)} for values in records],
         schema=pa.schema([pa.field(c, PYTHON_TYPE_TO_PA[columns[c]]) for c in columns]),
     )
-    if save_output:
-        pq.write_table(table, f"{output_name}.parquet")
-    return f"{output_name}.parquet", table
+    if output_filename:
+        pq.write_table(table, f"{output_filename}.parquet")
+    return f"{output_filename}.parquet", table
